refactor into utils
